[#12] 스프라이트 y좌표 Max 초과 짤림 구현
- byte 자료형 음수 표현 문제 short로 바꾸어 해결!

+++ src/render/decode_img.hpp
@@ -8,6 +8,6 @@
 void beginTV();
 void clearTV();
 void DecodeImg(byte image_byte, byte x_pos, byte y_pos);
-void DecodeSprite(byte image_byte, byte x_pos, byte y_pos);
+void DecodeSprite(byte image_byte, short x_pos, short y_pos);
 
 #endif
